refactor: APIの呼び出し処理を分割 ID指定で投稿を取得
