+++ src/store/postsSlice.ts
@@ -1,6 +1,7 @@
 import { createSlice, PayloadAction, createAsyncThunk } from '@reduxjs/toolkit';
 import type { RootState } from './store';
 import { Post } from '../domain/entities/Post';
+import { addPost, fetchPostById, fetchPosts } from '../api/postApi';
 
 // ステートの型
 interface PostsState {
@@ -15,28 +16,6 @@ const initialState: PostsState = {
   loading: false,
   error: null
 };
-
-export const addPost = createAsyncThunk(
-  'posts/addPost',
-  async (post: Omit<Post, 'id' | 'likes'>) => {
-    const response = await fetch(process.env.NEXT_PUBLIC_API_URL + '/api/posts', {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify(post),
-    });
-
-    if (!response.ok) throw new Error('投稿の追加に失敗しました');
-
-    return (await response.json()) as {"message": string};
-  },
-);
-
-// 非同期の投稿データ取得処理
-export const fetchPosts = createAsyncThunk('posts/fetchPosts', async () => {
-  const response = await fetch(process.env.NEXT_PUBLIC_API_URL + '/api/posts'); // Go の API にリクエスト
-  if (!response.ok) throw new Error('投稿の取得に失敗しました');
-  return (await response.json()) as Post[];
-});
 
 export const postsSlice = createSlice({
   name: 'posts',
@@ -69,6 +48,10 @@ export const postsSlice = createSlice({
       .addCase(fetchPosts.rejected, (state, action) => {
         state.loading = false;
         state.error = action.error.message ?? 'エラーが発生しました';
+      })
+      .addCase(fetchPostById.fulfilled, (state, action) => {
+        state.loading = false;
+        state.posts.push(action.payload);
       })
       .addCase(addPost.fulfilled, (state) => {
         state.loading = false;
